backend/requirements.txt
python-jose==3.3.0
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
gspread
oauth2client
pandas
axios
apscheduler

